+++ addOns/automation/src/main/java/org/zaproxy/addon/automation/jobs/ActiveScanPolicyJob.java
@@ -3,7 +3,7 @@
  *
  * ZAP is an HTTP/HTTPS proxy for assessing web application security.
  *
- * Copyright 2021 The ZAP Development Team
+ * Copyright 2024 The ZAP Development Team
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -19,15 +19,14 @@
  */
 package org.zaproxy.addon.automation.jobs;
 
-import java.util.ArrayList;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.concurrent.TimeUnit;
 import lombok.Getter;
 import lombok.Setter;
 import org.apache.commons.configuration.ConfigurationException;
-import org.apache.commons.lang3.StringUtils;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 import org.parosproxy.paros.Constant;
 import org.parosproxy.paros.control.Control;
 import org.parosproxy.paros.core.scanner.Plugin;
@@ -38,25 +37,18 @@ import org.zaproxy.addon.automation.AutomationData;
 import org.zaproxy.addon.automation.AutomationEnvironment;
 import org.zaproxy.addon.automation.AutomationJob;
 import org.zaproxy.addon.automation.AutomationProgress;
-import org.zaproxy.addon.automation.ContextWrapper;
-import org.zaproxy.addon.automation.JobResultData;
-import org.zaproxy.addon.automation.gui.ActiveScanJobDialog;
+import org.zaproxy.addon.automation.gui.ActiveScanPolicyJobDialog;
 import org.zaproxy.addon.automation.jobs.PolicyDefinition.Rule;
-import org.zaproxy.addon.commonlib.Constants;
-import org.zaproxy.zap.extension.ascan.ActiveScan;
 import org.zaproxy.zap.extension.ascan.ExtensionActiveScan;
 import org.zaproxy.zap.extension.ascan.ScanPolicy;
-import org.zaproxy.zap.model.Target;
-import org.zaproxy.zap.users.User;
 
-public class ActiveScanJob extends AutomationJob {
+public class ActiveScanPolicyJob extends AutomationJob {
 
-    public static final String JOB_NAME = "activeScan";
-    private static final String OPTIONS_METHOD_NAME = "getScannerParam";
+    public static final String JOB_NAME = "activeScan-policy";
 
-    private static final String PARAM_CONTEXT = "context";
-    private static final String PARAM_POLICY = "policy";
-    private static final String PARAM_USER = "user";
+    private static final Logger LOGGER = LogManager.getLogger(ActiveScanPolicyJob.class);
+
+    private static final String PARAM_NAME = "name";
 
     private ExtensionActiveScan extAScan;
 
@@ -64,18 +56,8 @@ public class ActiveScanJob extends AutomationJob {
     private PolicyDefinition policyDefinition = new PolicyDefinition();
     private Data data;
 
-    public ActiveScanJob() {
+    public ActiveScanPolicyJob() {
         data = new Data(this, this.parameters, this.policyDefinition);
-    }
-
-    @Override
-    public boolean supportsAlertTests() {
-        return true;
-    }
-
-    @Override
-    public String getKeyAlertTestsResultData() {
-        return ActiveScanJobResultData.KEY;
     }
 
     private ExtensionActiveScan getExtAScan() {
@@ -120,128 +102,48 @@ public class ActiveScanJob extends AutomationJob {
                     break;
             }
         }
-
-        this.verifyUser(this.getParameters().getUser(), progress);
-    }
-
-    @Override
-    public void applyParameters(AutomationProgress progress) {
-        JobUtils.applyObjectToObject(
-                this.parameters,
-                JobUtils.getJobOptions(this, progress),
-                this.getName(),
-                new String[] {PARAM_POLICY, PARAM_CONTEXT, PARAM_USER},
-                progress,
-                this.getPlan().getEnv());
     }
 
     @Override
     public Map<String, String> getCustomConfigParameters() {
         Map<String, String> map = super.getCustomConfigParameters();
-        map.put(PARAM_CONTEXT, "");
+        map.put(PARAM_NAME, "");
         return map;
     }
 
     @Override
-    public boolean supportsMonitorTests() {
-        return true;
+    public boolean applyCustomParameter(String name, String value) {
+        if (PARAM_NAME.equals(name)) {
+            return true;
+        }
+        return false;
     }
 
     @Override
     public void runJob(AutomationEnvironment env, AutomationProgress progress) {
+        try {
+            ScanPolicy scanPolicy = this.getScanPolicy(progress);
 
-        getExtAScan().setPanelSwitch(false);
-
-        ContextWrapper context;
-        if (StringUtils.isNotEmpty(this.getParameters().getContext())) {
-            context = env.getContextWrapper(this.getParameters().getContext());
-            if (context == null) {
-                progress.error(
+            List<String> allNames = getExtAScan().getPolicyManager().getAllPolicyNames();
+            if (allNames.contains(scanPolicy.getName())) {
+                progress.warn(
                         Constant.messages.getString(
-                                "automation.error.context.unknown",
-                                this.getParameters().getContext()));
+                                "automation.error.scanpolicy.exists",
+                                this.getName(),
+                                scanPolicy.getName()));
                 return;
             }
-        } else {
-            context = env.getDefaultContextWrapper();
+            getExtAScan().getPolicyManager().savePolicy(scanPolicy);
+        } catch (ConfigurationException e) {
+            progress.error(
+                    Constant.messages.getString("automation.dialog.error.misc", e.getMessage()));
+            LOGGER.error(e.getMessage(), e);
         }
-
-        Target target = new Target(context.getContext());
-        target.setRecurse(true);
-        List<Object> contextSpecificObjects = new ArrayList<>();
-        User user = this.getUser(this.getParameters().getUser(), progress);
-
-        ScanPolicy scanPolicy = null;
-        if (!StringUtils.isEmpty(this.getParameters().getPolicy())) {
-            try {
-                scanPolicy =
-                        this.getExtAScan()
-                                .getPolicyManager()
-                                .getPolicy(this.getParameters().getPolicy());
-            } catch (ConfigurationException e) {
-                // Error already raised above
-            }
-        } else {
-            scanPolicy = this.getScanPolicy(progress);
-        }
-        if (scanPolicy != null) {
-            contextSpecificObjects.add(scanPolicy);
-        }
-
-        int scanId = this.getExtAScan().startScan(target, user, contextSpecificObjects.toArray());
-
-        long endTime = Long.MAX_VALUE;
-        if (JobUtils.unBox(this.getParameters().getMaxScanDurationInMins()) > 0) {
-            // The active scan should stop, if it doesnt we will stop it (after a few seconds
-            // leeway)
-            endTime =
-                    System.currentTimeMillis()
-                            + TimeUnit.MINUTES.toMillis(
-                                    this.getParameters().getMaxScanDurationInMins())
-                            + TimeUnit.SECONDS.toMillis(5);
-        }
-
-        // Wait for the active scan to finish
-        ActiveScan scan;
-        boolean forceStop = false;
-
-        while (true) {
-            this.sleep(500);
-            scan = this.getExtAScan().getScan(scanId);
-            if (scan.isStopped()) {
-                break;
-            }
-            if (!this.runMonitorTests(progress) || System.currentTimeMillis() > endTime) {
-                forceStop = true;
-                break;
-            }
-        }
-        if (forceStop) {
-            this.getExtAScan().stopScan(scanId);
-            progress.info(Constant.messages.getString("automation.info.jobstopped", getType()));
-        }
-        progress.addJobResultData(createJobResultData(scanId));
-
-        getExtAScan().setPanelSwitch(true);
-    }
-
-    @Override
-    public List<JobResultData> getJobResultData() {
-        ActiveScan lastScan = this.getExtAScan().getLastScan();
-        if (lastScan != null) {
-            return createJobResultData(lastScan.getId());
-        }
-        return new ArrayList<>();
-    }
-
-    private List<JobResultData> createJobResultData(int scanId) {
-        List<JobResultData> list = new ArrayList<>();
-        list.add(new ActiveScanJobResultData(this.getName(), this.getExtAScan().getScan(scanId)));
-        return list;
     }
 
     protected ScanPolicy getScanPolicy(AutomationProgress progress) {
         ScanPolicy scanPolicy = new ScanPolicy();
+        scanPolicy.setName(this.getData().getParameters().getName());
 
         // Set default strength
         AttackStrength st =
@@ -313,33 +215,10 @@ public class ActiveScanJob extends AutomationJob {
     }
 
     @Override
-    public boolean isExcludeParam(String param) {
-        switch (param) {
-            case "allowAttackOnStart":
-            case "attackPolicy":
-            case "hostPerScan":
-            case "maxChartTimeInMins":
-            case "maxResultsToList":
-            case "maxScansInUI":
-            case "promptInAttackMode":
-            case "promptToClearFinishedScans":
-            case "rescanInAttackMode":
-            case "showAdvancedDialog":
-            case "targetParamsInjectable":
-            case "targetParamsEnabledRPC":
-                return true;
-            default:
-                return false;
-        }
-    }
-
-    @Override
     public String getSummary() {
-        String context = this.getParameters().getContext();
-        if (StringUtils.isEmpty(context)) {
-            context = Constant.messages.getString("automation.dialog.default");
-        }
-        return Constant.messages.getString("automation.dialog.ascan.summary", context);
+        return Constant.messages.getString(
+                "automation.dialog.ascanpolicy.summary",
+                JobUtils.unBox(this.getParameters().getName(), ""));
     }
 
     @Override
@@ -359,22 +238,22 @@ public class ActiveScanJob extends AutomationJob {
 
     @Override
     public Order getOrder() {
-        return Order.ATTACK;
+        return Order.CONFIGS;
     }
 
     @Override
     public Object getParamMethodObject() {
-        return this.getExtAScan();
+        return null;
     }
 
     @Override
     public String getParamMethodName() {
-        return OPTIONS_METHOD_NAME;
+        return null;
     }
 
     @Override
     public void showDialog() {
-        new ActiveScanJobDialog(this).setVisible(true);
+        new ActiveScanPolicyJobDialog(this).setVisible(true);
     }
 
     @Getter
@@ -392,26 +271,6 @@ public class ActiveScanJob extends AutomationJob {
     @Getter
     @Setter
     public static class Parameters extends AutomationData {
-        private String context = "";
-        private String user = "";
-        private String policy = "";
-        private Integer maxRuleDurationInMins = 0;
-        private Integer maxScanDurationInMins = 0;
-        private Boolean addQueryParam = false;
-        private String defaultPolicy = "";
-        private Integer delayInMs = 0;
-        private Boolean handleAntiCSRFTokens = true;
-        private Boolean injectPluginIdInHeader = false;
-        private Boolean scanHeadersAllRequests = false;
-        private Integer threadPerHost = Constants.getDefaultThreadCount();
-        private Integer maxAlertsPerRule = 0;
-
-        public Integer getThreadPerHost() {
-            if (JobUtils.unBox(threadPerHost) <= 0) {
-                // Don't return zero or less - this will cause problems
-                return null;
-            }
-            return threadPerHost;
-        }
+        private String name;
     }
 }